services/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
numpy
scikit-learn
pymongo
requests
yfinance
joblib
nsepython
xgboost
python-telegram-bot
